requirements.txt
fastapi==0.109.0
passlib==1.7.4
psycopg2_binary==2.9.9
pydantic==2.5.3
python-dotenv==1.0.0
python_jose==3.3.0
SQLAlchemy==1.4.0
uvicorn==0.25.0
